requirements.txt
langchain
langchain-community
langchain_groq
streamlit
python-dotenv
fastapi
langserve